+++ R/reporter-list.r
@@ -4,7 +4,7 @@ NULL
 
 setOldClass('proc_time')
 
-#' List reporter: gather all test results along with elapsed time and 
+#' List reporter: gather all test results along with elapsed time and
 #' file information.
 #'
 #' This reporter gathers all results, adding additional information such as
@@ -21,7 +21,7 @@ ListReporter <- setRefClass("ListReporter", contains = "Reporter",
     file = 'character',
     results = 'list',
     current_test_results = 'list'),
-    
+
   methods = list(
     ### overriden methods from Reporter
     start_reporter = function(...) {
@@ -29,22 +29,22 @@ ListReporter <- setRefClass("ListReporter", contains = "Reporter",
       results <<- list()
       current_test_results <<- list()
     },
-    
+
     start_test = function(desc) {
       callSuper(desc)
       current_test_results <<- list()
       start_test_time <<- proc.time()
     },
-    
+
     end_test = function() {
       el <- as.double(proc.time() - start_test_time)
       fname <- if (length(file)) file else ''
-      test_info <- list(file = fname, context = context, test = test, 
-        user = el[1], system = el[2], real = el[3], 
+      test_info <- list(file = fname, context = context, test = test,
+        user = el[1], system = el[2], real = el[3],
         results = current_test_results)
       results <<- c(results, list(test_info))
       current_test_results <<- list()
-      
+
       callSuper() # at the end because it resets the test name
     },
 
@@ -64,7 +64,6 @@ ListReporter <- setRefClass("ListReporter", contains = "Reporter",
 
   )
 )
-