+++ R/reporter-summary.r
@@ -80,31 +80,15 @@ SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
         cat("\n")
         if (has_tests && sample(10, 1) == 1 && show_praise) {
           cat(colourise(sample(.praise, 1), "passed"), "\n")
+        } else {
+          cat(colourise("DONE", "passed"), "\n")
         }
       } else {
-        label <- labels[seq_len(n)]
-        type <- ifelse(sapply(failures, "[[", "error"), "Error", "Failure")
-        tests <- vapply(failures, "[[", "test", FUN.VALUE = character(1))
-
-        location <- vapply(failures, function(x) {
-          ref <- x$srcref
-          if ( is.null(ref) ) {
-            ''
-          } else {
-            paste0('(@', attr(ref, 'srcfile')$filename, '#', ref[1], ')')
-          }
-        }, '')
-        header <- paste0(label, ". ", type, location, ": ", tests, " ")
-
-        linewidth <- ifelse(nchar(header) > getOption("width"),0,getOption("width") - nchar(header))
-        line <- charrep("-", linewidth )
-
-        message <- vapply(failures, "[[", "failure_msg", FUN.VALUE = character(1))
-
-        reports <- paste0(
-          colourise(header, "error"), line, "\n",
-          message, "\n")
-        cat("\n", reports, sep = "\n")
+        cat("\n")
+        reports <- vapply(seq_len(n), function(i) {
+          failure_summary(failures[[i]], labels[i])
+        }, character(1))
+        cat(paste(reports, collapse = "\n\n"), "\n", sep = "")
       }
     }
   )